+++ internal/cli/root.go
@@ -27,6 +27,7 @@ func NewRootCmd(stdout, stderr io.Writer) *cobra.Command {
 	root.AddCommand(newValidateCmd())
 	root.AddCommand(newApplyCmd())
 	root.AddCommand(newDiffCmd())
+	root.AddCommand(newBenchCmd())
 	root.AddCommand(newDoctorCmd())
 	root.AddCommand(newRemoveCmd())
 	root.AddCommand(newUninstallCmd())